refactor: move mock data to lib/data
