+++ app/(pages)/upload/client.tsx
@@ -25,6 +25,7 @@ import {
 } from "@/components/ui/table";
 
 import { ImportHistory } from "@/lib/types";
+import { importHistory as mockImportHistory } from "@/lib/data/import-data";
 
 // Create a client component that uses useSearchParams
 function UploadContent() {
@@ -72,44 +73,8 @@ function UploadContent() {
     setManualImportType(event.target.value);
   };
 
-  // Mock import history data
-  const [importHistory] = React.useState<ImportHistory[]>([
-    {
-      id: "IMP2024001",
-      date: "2024-12-02 09:15",
-      fileName: "1202出貨紀錄.csv",
-      recordCount: 50,
-      status: "部份成功",
-    },
-    {
-      id: "IMP2024002",
-      date: "2024-12-05 09:15",
-      fileName: "1205出貨紀錄.csv",
-      recordCount: 150,
-      status: "失敗",
-    },
-    {
-      id: "IMP2024003",
-      date: "2024-12-30 09:15",
-      fileName: "1230出貨紀錄.csv",
-      recordCount: 250,
-      status: "成功",
-    },
-    {
-      id: "IMP2025001",
-      date: "2025-01-22 09:15",
-      fileName: "0122出貨紀錄.csv",
-      recordCount: 10,
-      status: "成功",
-    },
-    {
-      id: "IMP2025002",
-      date: "2025-02-24 09:15",
-      fileName: "0224出貨紀錄.csv",
-      recordCount: 150,
-      status: "成功",
-    },
-  ]);
+  // Import history data from centralized location
+  const [importHistory] = React.useState<ImportHistory[]>(mockImportHistory);
 
   // Validation handlers
