+++ lib/data/import-data.ts
@@ -1,0 +1,39 @@
+import { ImportHistory } from '../types';
+
+export const importHistory: ImportHistory[] = [
+  {
+    id: "IMP2024001",
+    date: "2024-12-02 09:15",
+    fileName: "1202出貨紀錄.csv",
+    recordCount: 50,
+    status: "部份成功",
+  },
+  {
+    id: "IMP2024002",
+    date: "2024-12-05 09:15",
+    fileName: "1205出貨紀錄.csv",
+    recordCount: 150,
+    status: "失敗",
+  },
+  {
+    id: "IMP2024003",
+    date: "2024-12-30 09:15",
+    fileName: "1230出貨紀錄.csv",
+    recordCount: 250,
+    status: "成功",
+  },
+  {
+    id: "IMP2025001",
+    date: "2025-01-22 09:15",
+    fileName: "0122出貨紀錄.csv",
+    recordCount: 10,
+    status: "成功",
+  },
+  {
+    id: "IMP2025002",
+    date: "2025-02-24 09:15",
+    fileName: "0224出貨紀錄.csv",
+    recordCount: 150,
+    status: "成功",
+  },
+];
